// common/elements/Button/interfaces.ts
export interface ButtonProps {
    title?: any;
    // textcolor?: string;
    // backgroundcolor?: string;
    // backgroundhovercolor?: string;
    className?: string;
    display?: string;
    width?: string;
    height?: string;
    padding?: string;
    fontSize?: string;
    icon?: string;
    disabled?: boolean;
    onClick?: () => void;
}